deleted extra files, update InitDependenciesCommand
