--- SpaceBattle/ActionCommand.cs
@@ -1,14 +1,0 @@
-﻿namespace SpaceBattle;
-
-public class ActionCommand : ICommand
-{
-    private readonly Action _action;
-    public ActionCommand(Action action)
-    {
-        _action = action;
-    }
-    public void Execute()
-    {
-        _action();
-    }
-}
